Completed the core aspect of the quiz. Forgot to add the multipliers for the time. Not enough time to reset the timer...for now. Didn't have time to display the leaderboards.

+++ FlashQuizHero/ResultsViewController.h
@@ -10,6 +10,6 @@
 
 @interface ResultsViewController : UIViewController
 
-- (void)loadScore:(int) score;
+- (void)loadScore:(float) score;
 
 @end
